정렬 및 기타 기능

+++ src/main/java/com/spring/mvc/web/michelin/repository/michelinMapper.java
@@ -22,7 +22,7 @@ public interface michelinMapper {
     List<Michelin> starFindAll();
 
     //글 수정
-    void modify(int board_num, String title, String content, String writer, int star);
+    void modify(int board_num, String title, String content, String writer, double star);
 
 
 }
